refactor: usePathname 대신 use를 이용하여 params 추출하는 방식으로 수정 #52
