+++ src/app/concert/[id]/_shared/components/concert-info/concert-info.tsx
@@ -10,7 +10,7 @@ import {
   TICKET_SITE_LABEL_MAP,
 } from '@/shared/constants/ticketsite-url';
 import { TicketReservationSite, Concert } from '@/shared/types';
-import calculateDday from '@/shared/utils/caclueate-dday';
+import calculateDday from '@/shared/utils/calculate-dday';
 import formatDate from '@/shared/utils/services/format-date';
 
 import styles from './concert-info.module.scss';
